+++ src/main/java/shop/bookbom/front/domain/user/adapter/impl/UserAdapterImpl.java
@@ -16,16 +16,20 @@ import org.springframework.web.util.UriComponentsBuilder;
 import shop.bookbom.front.common.CommonPage;
 import shop.bookbom.front.common.CommonResponse;
 import shop.bookbom.front.common.exception.RestTemplateException;
-import shop.bookbom.front.domain.member.dto.request.OrderDateCondition;
 import shop.bookbom.front.domain.order.dto.response.OrderInfoResponse;
 import shop.bookbom.front.domain.user.adapter.UserAdapter;
+import shop.bookbom.front.domain.user.dto.OrderDateCondition;
 import shop.bookbom.front.domain.user.dto.SignUpDto;
 import shop.bookbom.front.domain.user.dto.response.EmailCheckResponse;
+import shop.bookbom.front.domain.user.dto.response.UserInfoResponse;
 
 @Component
 @RequiredArgsConstructor
 public class UserAdapterImpl implements UserAdapter {
     private static final ParameterizedTypeReference<CommonResponse<CommonPage<OrderInfoResponse>>> ORDER_INFO_RESPONSE =
+            new ParameterizedTypeReference<>() {
+            };
+    private static final ParameterizedTypeReference<CommonResponse<UserInfoResponse>> MEMBER_INFO =
             new ParameterizedTypeReference<>() {
             };
     private static final ParameterizedTypeReference<CommonResponse<EmailCheckResponse>> EMAIL_CHECK_RESPONSE =
@@ -64,6 +68,26 @@ public class UserAdapterImpl implements UserAdapter {
             throw new RestTemplateException();
         }
         return Objects.requireNonNull(response).getResult();
+    }
+
+    @Override
+    public UserInfoResponse getUserInfo() {
+        HttpHeaders httpHeaders = new HttpHeaders();
+        httpHeaders.setContentType(MediaType.APPLICATION_JSON);
+        HttpEntity<Void> requestEntity = new HttpEntity<>(httpHeaders);
+
+        String url = UriComponentsBuilder.fromHttpUrl(gatewayUrl + "/shop/users/my-page")
+                .toUriString();
+
+        CommonResponse<UserInfoResponse> response = restTemplate.exchange(
+                url,
+                HttpMethod.GET,
+                requestEntity,
+                MEMBER_INFO).getBody();
+        if (response == null || !response.getHeader().isSuccessful()) {
+            throw new RestTemplateException();
+        }
+        return response.getResult();
     }
