+++ src/main/java/shop/bookbom/front/domain/user/dto/OrderDateCondition.java
@@ -1,4 +1,4 @@
-package shop.bookbom.front.domain.member.dto.request;
+package shop.bookbom.front.domain.user.dto;
 
 import java.time.LocalDate;
 import lombok.AllArgsConstructor;
