+++ src/main/java/shop/bookbom/front/domain/user/service/impl/UserServiceImpl.java
@@ -5,12 +5,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import shop.bookbom.front.domain.member.dto.request.OrderDateCondition;
 import shop.bookbom.front.domain.order.dto.response.OrderInfoResponse;
 import shop.bookbom.front.domain.user.adapter.UserAdapter;
+import shop.bookbom.front.domain.user.dto.OrderDateCondition;
 import shop.bookbom.front.domain.user.dto.SignUpDto;
 import shop.bookbom.front.domain.user.dto.request.SignUpRequest;
 import shop.bookbom.front.domain.user.dto.response.EmailCheckResponse;
+import shop.bookbom.front.domain.user.dto.response.UserInfoResponse;
 import shop.bookbom.front.domain.user.service.UserService;
 
 @Service
@@ -35,5 +36,10 @@ public class UserServiceImpl implements UserService {
         // todo 비밀번호 암호화
         String encodePassword = password;
         userAdapter.signUp(SignUpDto.of(request, encodePassword));
+    }
+
+    @Override
+    public UserInfoResponse getUserInfo() {
+        return userAdapter.getUserInfo();
     }
 }
