+++ src/main/resources/templates/fragments/header.html
@@ -34,7 +34,7 @@
                 <div class="d-none d-lg-flex align-items-end">
                     <ul class="d-flex justify-content-end list-unstyled m-0">
                         <li>
-                            <a href="account.html" class="mx-3">
+                            <a class="mx-3" href="/users/my-page">
                                 <iconify-icon icon="healthicons:person" class="fs-4"></iconify-icon>
                             </a>
                         </li>
