+++ src/main/resources/templates/page/user/my-page.html
@@ -15,9 +15,9 @@
 <div class="container mt-5" layout:fragment="content">
     <div class="row justify-content-center mb-3">
         <div class="col-md-6 text-center">
-            <span class="fw-bold fs-2" th:text="${member.nickname}">냥냥</span><span class="fs-4">님</span>
+            <span class="fw-bold fs-2" th:text="${user.nickname}">냥냥</span><span class="fs-4">님</span>
             <div class="member-link" th:data-href="@{/rank-policy}">
-                <span class="fs-5" th:text="${member.rank}">PLATINUM</span>
+                <span class="fs-5" th:text="${user.rank}">PLATINUM</span>
             </div>
         </div>
     </div>
@@ -25,10 +25,10 @@
         <div class="col-md-3">
             <span class="fw-bold fs-4">쇼핑</span>
             <ul class="list-group mb-3">
-                <li class="list-group-item member-link" th:data-href="@{/orders}">주문 내역</li>
-                <li class="list-group-item member-link" th:data-href="@{/wishes}">찜 목록</li>
-                <li class="list-group-item member-link" th:data-href="@{/coupons}">쿠폰함</li>
-                <li class="list-group-item member-link" th:data-href="@{/points}">포인트 적립/사용 목록</li>
+                <li class="list-group-item member-link" th:data-href="@{/users/orders}">주문 내역</li>
+                <li class="list-group-item member-link" th:data-href="@{/users/wishes}">찜 목록</li>
+                <li class="list-group-item member-link" th:data-href="@{/users/coupons}">쿠폰함</li>
+                <li class="list-group-item member-link" th:data-href="@{/users/point-history}">포인트 적립/사용 목록</li>
             </ul>
             <!-- 회원 정보 -->
             <span class="fw-bold fs-4">회원 정보</span>
@@ -60,7 +60,7 @@
                 </tr>
                 </thead>
                 <tbody>
-                <tr class="order-detail" th:data-order-id="${order.id}" th:each="order : ${member.lastOrders}">
+                <tr class="order-detail" th:data-order-id="${order.id}" th:each="order : ${user.lastOrders}">
                     <td th:text="${order.orderDate}">2024-04-25</td>
                     <td th:text="${order.orderNumber}">ORD123456</td>
                     <td th:text="${order.name}">상품명</td>
@@ -72,25 +72,25 @@
             <div class="row mt-4 mx-4 justify-content-between my-2">
                 <div class="col-md-2 my-2">
                     <div class="card custom-card" style="border-radius: 25px">
-                        <div class="card-body member-link" th:data-href="@{/coupons}">
+                        <div class="card-body member-link" th:data-href="@{/users/coupons}">
                             <h5 class="card-title text-center fw-bold">쿠폰함</h5>
-                            <p class="card-text text-end" th:text="${member.couponCount} + '장'">3장</p>
+                            <p class="card-text text-end" th:text="${user.couponCount} + '장'">3장</p>
                         </div>
                     </div>
                 </div>
                 <div class="col-md-2 my-2">
                     <div class="card custom-card" style="border-radius: 25px">
-                        <div class="card-body member-link" th:data-href="@{/wishes}">
+                        <div class="card-body member-link" th:data-href="@{/users/wishes}">
                             <h5 class="card-title text-center fw-bold">찜 목록</h5>
-                            <p class="card-text text-end" th:text="${member.wishCount} + '개'">10개</p>
+                            <p class="card-text text-end" th:text="${user.wishCount} + '개'">10개</p>
                         </div>
                     </div>
                 </div>
                 <div class="col-md-2 my-2">
                     <div class="card custom-card" style="border-radius: 25px">
-                        <div class="card-body member-link" th:data-href="@{/points}">
+                        <div class="card-body member-link" th:data-href="@{/users/point-history}">
                             <h5 class="card-title text-center fw-bold">포인트</h5>
-                            <p class="card-text text-end" th:text="${member.point} + 'p'">2500p</p>
+                            <p class="card-text text-end" th:text="${user.point} + 'p'">2500p</p>
                         </div>
                     </div>
                 </div>
